refactor(editor): Move setup templates to templates feature
