+++ packages/frontend/editor-ui/src/components/Node/NodeCreator/utils.ts
@@ -40,7 +40,10 @@ import { useNodeTypesStore } from '@/stores/nodeTypes.store';
 import { SEND_AND_WAIT_OPERATION } from 'n8n-workflow';
 import type { NodeIconSource } from '../../../utils/nodeIcon';
 import type { CommunityNodeDetails, ViewStack } from './composables/useViewStacks';
-import { PrebuiltAgentTemplates, SampleTemplates } from '@/utils/templates/workflowSamples';
+import {
+	PrebuiltAgentTemplates,
+	SampleTemplates,
+} from '@/features/templates/utils/workflowSamples';
 
 const COMMUNITY_NODE_TYPE_PREVIEW_TOKEN = '-preview';
