+++ packages/frontend/editor-ui/src/composables/useCalloutHelpers.test.ts
@@ -1,7 +1,10 @@
 import { useCalloutHelpers } from '@/composables/useCalloutHelpers';
 import { updateCurrentUserSettings } from '@n8n/rest-api-client/api/users';
 import { createTestingPinia } from '@pinia/testing';
-import { PrebuiltAgentTemplates, SampleTemplates } from '@/utils/templates/workflowSamples';
+import {
+	PrebuiltAgentTemplates,
+	SampleTemplates,
+} from '@/features/templates/utils/workflowSamples';
 import { useNDVStore } from '@/stores/ndv.store';
 import { mockedStore } from '@/__tests__/utils';
 import { NODE_CREATOR_OPEN_SOURCES, VIEWS } from '@/constants';
