+++ packages/frontend/editor-ui/src/composables/useCalloutHelpers.ts
@@ -26,7 +26,7 @@ import {
 	isPrebuiltAgentTemplateId,
 	isTutorialTemplateId,
 	SampleTemplates,
-} from '@/utils/templates/workflowSamples';
+} from '@/features/templates/utils/workflowSamples';
 import type { INodeCreateElement, OpenTemplateElement } from '@/Interface';
 import { useUIStore } from '@/stores/ui.store';
 import { useProjectsStore } from '@/stores/projects.store';
