+++ packages/frontend/editor-ui/src/features/templates/__tests__/setupTemplate.store.testData.ts
