+++ packages/frontend/editor-ui/src/features/templates/components/IconSuccess.vue
