+++ packages/frontend/editor-ui/src/features/templates/components/SetupWorkflowCredentialsButton.vue
