+++ packages/frontend/editor-ui/src/features/templates/utils/samples/agents/calendar-agent.json
