+++ packages/frontend/editor-ui/src/features/templates/utils/samples/agents/email_triage_agent_with_gmail.json
