+++ packages/frontend/editor-ui/src/features/templates/utils/samples/agents/joke_agent_with_http_tool.json
