+++ packages/frontend/editor-ui/src/features/templates/utils/samples/agents/knowledge_store_agent_with_google_drive.json
