+++ packages/frontend/editor-ui/src/features/templates/utils/samples/agents/voice-agent.json
