+++ packages/frontend/editor-ui/src/features/templates/utils/samples/easy_ai_starter.json
