+++ packages/frontend/editor-ui/src/features/templates/utils/samples/rag_starter.json
