+++ packages/frontend/editor-ui/src/features/templates/utils/samples/tutorial/api_fundamentals.json
