+++ packages/frontend/editor-ui/src/features/templates/utils/samples/tutorial/build_your_first_ai_agent.json
