+++ packages/frontend/editor-ui/src/features/templates/utils/samples/tutorial/json_basics.json
