+++ packages/frontend/editor-ui/src/features/templates/utils/templateActions.ts
@@ -7,8 +7,8 @@ import type { useRootStore } from '@n8n/stores/useRootStore';
 import type { useWorkflowsStore } from '@/stores/workflows.store';
 import { getNodesWithNormalizedPosition } from '@/utils/nodeViewUtils';
 import type { NodeTypeProvider } from '@/utils/nodeTypes/nodeTypeTransforms';
-import type { TemplateCredentialKey } from '@/utils/templates/templateTransforms';
-import { replaceAllTemplateNodeCredentials } from '@/utils/templates/templateTransforms';
+import type { TemplateCredentialKey } from './templateTransforms';
+import { replaceAllTemplateNodeCredentials } from './templateTransforms';
 import type { INodeCredentialsDetails } from 'n8n-workflow';
 import type { RouteLocationRaw, Router } from 'vue-router';
 import type { TemplatesStore } from '@/features/templates/templates.store';
