+++ packages/frontend/editor-ui/src/features/templates/utils/typeGuards.ts
