+++ packages/frontend/editor-ui/src/features/templates/views/TemplatesCollectionView.vue
@@ -6,10 +6,10 @@ import TemplatesView from './TemplatesView.vue';
 import type { ITemplatesWorkflow } from '@n8n/rest-api-client/api/templates';
 import { VIEWS } from '@/constants';
 import { useTemplatesStore } from '@/features/templates/templates.store';
-import { useTemplateWorkflow } from '@/utils/templates/templateActions';
+import { useTemplateWorkflow } from '@/features/templates/utils/templateActions';
 import { useExternalHooks } from '@/composables/useExternalHooks';
 import { useNodeTypesStore } from '@/stores/nodeTypes.store';
-import { isFullTemplatesCollection } from '@/utils/templates/typeGuards';
+import { isFullTemplatesCollection } from '@/features/templates/utils/typeGuards';
 import { useRoute, useRouter } from 'vue-router';
 import { useTelemetry } from '@/composables/useTelemetry';
 import { useDocumentTitle } from '@/composables/useDocumentTitle';
